Move dispatcher and mobile blueprints into new package
